Platform that allows users to type to the lyrics of a song of their choice for entertainment and typing practice. Uses an SQL database to keep track of user submissions in a leaderboard. Uses Genius and Spotify APIs to search for a song based on the user’s search query and acquire lyrics.
Platform that allows users to type to the lyrics of a song of their choice for entertainment and typing practice. Uses an SQL database to keep track of user submissions in a leaderboard. Uses Genius and Spotify APIs to search for a song based on the user’s search query and acquire lyrics.